fix(chatroom): prevent server from crashing if a user that is not a member of the chatroom leaves

+++ chatara-backend/chatroom.js
@@ -21,14 +21,14 @@ router.get('/:id', authenticate, async (req, res) => {
 
 router.post('/leave/:chatroomId', authenticate, async (req, res) => {
     const chatroomId = req.params.chatroomId
-    if (!chatroomId) return res.status(400).json({ message: 'Missing chatroom Id' })
+    if (!chatroomId || typeof (chatroomId) !== 'string' || chatroomId.trim() === '') return res.status(400).json({ message: 'Missing or invalid chatroom Id' })
     const userId = req.id
     if (!userId) return res.status(401).json({ message: 'Unauthorized' })
 
     try {
         const selectQuery = `SELECT u.username FROM user_tbl u INNER JOIN participant_tbl p ON u.id = p.user_id WHERE u.id = ? AND p.chatroom_id = ?`
-        const [user] = await pool.execute(selectQuery, [userId, chatroomId])
-        if (!user[0].username) return res.status(400).json({ message: `User is not a member of the chatroom` })
+        const [rows] = await pool.execute(selectQuery, [userId, chatroomId])
+        if (rows.length === 0) return res.status(400).json({ message: `User is not a member of the chatroom` })
         const deleteQuery = `DELETE FROM participant_tbl WHERE user_id = ? AND chatroom_id = ?`
         await pool.execute(deleteQuery, [userId, chatroomId])
         res.status(200).json({ message: 'Left chatroom' })
